requirements.txt
python=3.8.8
pytorch=1.7.1
torchvision=0.8.2
numpy=1.23.5
scipy=1.10.0
opencv-python=4.7.0
matplotlib=3.2.2
tqdm=4.64.1
h5py=3.7.0
imageio=2.26.0
pandas=1.2.3
nibabel=5.0.1
scikit-learn=1.2.1
tifffile=2020.10.1
lightning-bolts=0.5.0
pytorch-lightning=1.5.8
scikit-image=0.19.3
filelock=3.9.0
seaborn=0.12.2


